remove plans service
