+++ app/components/billing/account.js
@@ -5,7 +5,6 @@ import { reads, empty, bool, not, and } from '@ember/object/computed';
 
 export default Component.extend({
   store: service(),
-  plan: service(),
   accounts: service(),
 
   account: null,
@@ -32,6 +31,6 @@ export default Component.extend({
 
   init() {
     this._super(...arguments);
-    this.plan.fetchPlans.perform(this.account);
+    this.account.fetchPlans.perform();
   }
 });
